postgresql

+++ src/main/java/parking/domain/models/Usuarios.java
@@ -6,7 +6,7 @@ import jakarta.persistence.Id;
 public class Usuarios{
     @Id
     @GeneratedValue
-    private Long id;
+    private Integer id;
 
     private String nombre;
     private String email;
@@ -16,7 +16,7 @@ public class Usuarios{
 
     public Usuarios() {
     }
-    public  Usuarios(Long id, String nombre, String email, String dni, Integer telefono, String rol) {
+    public  Usuarios(Integer id, String nombre, String email, String dni, Integer telefono, String rol) {
         this.id = id;
         this.nombre = nombre;
         this.email = email;
